fix(mem): add explicit uint8_t casts to fix GCC -Wconversion errors
Integer promotion in (uint8_t + uint8_t) % uint8_t produces int,
which triggers -Werror=conversion when assigned back to uint8_t
on GCC (Ubuntu 24.04, -std=c++20 with -Wconversion). Add explicit
static_cast<uint8_t> around bin index and start_bin computations.

+++ src/mem/slab_cache.cpp
@@ -52,10 +52,12 @@ void* SlabCache::allocate(ActorId owner) noexcept {
     } else if (strategy_ == AllocationStrategy::kSegregatedFit) {
         // Segregated: round-robin search across bins, one block per bin
         for (uint8_t i = 0; i < kNumSegregatedBins; ++i) {
-            uint8_t bin_idx = (start_bin_ + i) % kNumSegregatedBins;
+            uint8_t bin_idx =
+                static_cast<uint8_t>((start_bin_ + i) % kNumSegregatedBins);
             block = bins_[bin_idx].pop();
             if (block) {
-                start_bin_ = (bin_idx + 1) % kNumSegregatedBins;
+                start_bin_ =
+                    static_cast<uint8_t>((bin_idx + 1) % kNumSegregatedBins);
                 break;
             }
         }
